Move production types to the EssentialFeedII production module
