+++ EssentialFeedII/Feed Presentation/FeedPresenter.swift
@@ -1,0 +1,95 @@
+//
+//  FeedPresenter.swift
+//  EssentialFeedII
+//
+//  Created by Julius on 18/09/2023.
+//
+
+import Foundation
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
+
+public protocol FeedView {
+    func display(_ viewModel: FeedViewModel) // happens in the uitableview
+}
+
+public struct LocalFeedImage: Equatable {
+    public let id: UUID
+    public let description: String?
+    public let location: String?
+    public let url: URL
+    
+    public init(id: UUID, description: String?, location: String?, url: URL) {
+        self.id = id
+        self.description = description
+        self.location = location
+        self.url = url
+    }
+}
+
+public struct FeedLoadingViewModel {
+    public let isLoading: Bool
+}
+
+public protocol FeedLoadingView {
+    func display(_ viewModel: FeedLoadingViewModel)// happens in the uirefreshcontrol
+}
+
+public struct FeedErrorViewModel {
+    public let message: String?
+    
+    static var noError: FeedErrorViewModel {
+        return FeedErrorViewModel(message: nil)
+    }
+    
+    static func error(message: String) -> FeedErrorViewModel {
+        return FeedErrorViewModel(message: message)
+    }
+}
+
+public protocol FeedErrorView {
+    func display(_ viewModel: FeedErrorViewModel)
+}
+
+public final class FeedPresenter {
+    private let feedView: FeedView
+    private let feedLoadingView: FeedLoadingView
+    private let errorView: FeedErrorView
+    
+    public init(feedView: FeedView, feedLoadingView: FeedLoadingView, errorView: FeedErrorView) {
+        self.feedView = feedView
+        self.feedLoadingView = feedLoadingView
+        self.errorView = errorView
+    }
+    
+    public static var title: String {
+        return NSLocalizedString("FEED_VIEW_TITLE",
+                                 tableName: "Feed",
+                                 bundle: Bundle(for: FeedPresenter.self),
+                                 comment: "Title for the feed view")
+    }
+    
+    private var feedLoadError: String {
+        return NSLocalizedString("FEED_VIEW_CONNECTION_ERROR",
+                                 tableName: "Feed",
+                                 bundle: Bundle(for: FeedPresenter.self),
+                                 comment: "Error message displayed when we can't load the image feed from the server")
+    }
+    
+    public func didStartLoadingFeed() {
+        errorView.display(.noError)
+        feedLoadingView.display(FeedLoadingViewModel(isLoading: true))
+    }
+    
+    public func didFinishLoadingFeed(with feed: [FeedImage]) {
+        feedView.display(FeedViewModel(feed: feed))
+        feedLoadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+    
+    public func didFinishLoading(with error: Error) {
+        errorView.display(.error(message: feedLoadError))
+        feedLoadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+}
