+++ EssentialFeedIITests/Feed Presentation/FeedPresenterTests.swift
@@ -8,93 +8,6 @@
 import XCTest
 import EssentialFeedII
 
-struct FeedViewModel {
-    let feed: [FeedImage]
-}
-
-protocol FeedView {
-    func display(_ viewModel: FeedViewModel) // happens in the uitableview
-}
-
-public struct LocalFeedImage: Equatable {
-    public let id: UUID
-    public let description: String?
-    public let location: String?
-    public let url: URL
-    
-    public init(id: UUID, description: String?, location: String?, url: URL) {
-        self.id = id
-        self.description = description
-        self.location = location
-        self.url = url
-    }
-}
-
-struct FeedLoadingViewModel {
-    let isLoading: Bool
-}
-
-protocol FeedLoadingView {
-    func display(_ viewModel: FeedLoadingViewModel)// happens in the uirefreshcontrol
-}
-
-struct FeedErrorViewModel {
-    let message: String?
-    
-    static var noError: FeedErrorViewModel {
-        return FeedErrorViewModel(message: nil)
-    }
-    
-    static func error(message: String) -> FeedErrorViewModel {
-        return FeedErrorViewModel(message: message)
-    }
-}
-
-protocol FeedErrorView {
-    func display(_ viewModel: FeedErrorViewModel)
-}
-
-class FeedPresenter {
-    private let feedView: FeedView
-    private let feedLoadingView: FeedLoadingView
-    private let errorView: FeedErrorView
-    
-    init(feedView: FeedView, feedLoadingView: FeedLoadingView, errorView: FeedErrorView) {
-        self.feedView = feedView
-        self.feedLoadingView = feedLoadingView
-        self.errorView = errorView
-    }
-    
-    static var title: String {
-        return NSLocalizedString("FEED_VIEW_TITLE",
-                                 tableName: "Feed",
-                                 bundle: Bundle(for: FeedPresenter.self),
-                                 comment: "Title for the feed view")
-    }
-    
-    private var feedLoadError: String {
-        return NSLocalizedString("FEED_VIEW_CONNECTION_ERROR",
-                                 tableName: "Feed",
-                                 bundle: Bundle(for: FeedPresenter.self),
-                                 comment: "Error message displayed when we can't load the image feed from the server")
-    }
-    
-    func didStartLoadingFeed() {
-        errorView.display(.noError)
-        feedLoadingView.display(FeedLoadingViewModel(isLoading: true))
-    }
-    
-    func didFinishLoadingFeed(with feed: [FeedImage]) {
-        feedView.display(FeedViewModel(feed: feed))
-        feedLoadingView.display(FeedLoadingViewModel(isLoading: false))
-    }
-    
-    func didFinishLoading(with error: Error) {
-        errorView.display(.error(message: feedLoadError))
-        feedLoadingView.display(FeedLoadingViewModel(isLoading: false))
-    }
-}
-
 final class FeedPresenterTests: XCTestCase {
     
     func test_title_isLocalized() {
